Adding final changes for Fall '24
If any bug is not caught here, then it will be caught in the Spring 2025 update branch or PR.
