+++ .astro/settings.json
@@ -1,5 +1,5 @@
 {
 	"_variables": {
-		"lastUpdateCheck": 1723428618787
+		"lastUpdateCheck": 1747265742551
 	}
 }
