+++ package.json
@@ -17,7 +17,7 @@
     "astro": "astro"
   },
   "dependencies": {
-    "astro": "^4.13.3"
+    "astro": "^5.7.13"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^8.0.1",
