+++ src/pages/about.astro
@@ -39,24 +39,25 @@ import Hero from "../components/Hero.astro";
 						and Spain. My journey in learning continues as I delve
 						into Java, Assembly, C, and web programming.
 						Academically, I have successfully completed courses like
-						M 408C (Calculus I), M 340L (Matrices and Matrix Calculations) 
-						and ECO 304K/L (Micro and
-						Macroeconomics), which have broadened my analytical and
-						economic understanding. Outside the classroom, I am
-						actively involved in software development, with numerous
-						projects showcased on my GitHub profile. My leadership
-						qualities have been honed through roles such as team
-						leader and CEO of my own startup. Beyond my professional
-						life, I am passionate about soccer, video gaming, and
-						programming for fun. These hobbies not only provide a
-						creative outlet but also keep me grounded and connected
-						to my community. My aspiration is to inspire change
-						within my community, emphasizing that education is a key
-						catalyst for progress. I believe in seizing
-						opportunities and making the most of our interconnected
-						world. With a commitment to continuous learning and
-						development, I aim to contribute significantly to the
-						field of software engineering and beyond.
+						M 408C/D (Calculus I and II), M 340L (Matrices and
+						Matrix Calculations), ECO 304K/L (Micro and
+						Macroeconomics), and PHY 303K (Engineering Physics I),
+						which have broadened my analytical and economic
+						understanding. Outside the classroom, I am actively
+						involved in software development, with numerous projects
+						showcased on my GitHub profile. My leadership qualities
+						have been honed through roles such as team leader and
+						CEO of my own startup. Beyond my professional life, I am
+						passionate about soccer, video gaming, and programming
+						for fun. These hobbies not only provide a creative
+						outlet but also keep me grounded and connected to my
+						community. My aspiration is to inspire change within my
+						community, emphasizing that education is a key catalyst
+						for progress. I believe in seizing opportunities and
+						making the most of our interconnected world. With a
+						commitment to continuous learning and development, I aim
+						to contribute significantly to the field of software
+						engineering and beyond.
 					</p>
 				</div>
 				<section>
@@ -66,7 +67,10 @@ import Hero from "../components/Hero.astro";
 							<b>Studying at the University of Texas at Austin</b>
 						</p>
 						<p><b><i>Economics, Bachelor of Science</i></b></p>
-						<p>with a certificate in Computer Science</p><br>
+						<p>
+							with a certificate in Computer Science and
+							Engineering
+						</p><br />
 						<p><i>Expected Graduation Date: May 2027</i></p>
 					</div>
 				</section>
@@ -74,9 +78,10 @@ import Hero from "../components/Hero.astro";
 					<h2 class="section-title">Skills</h2>
 					<div class="content">
 						<p>
-							Intermediate Python, Intermediate Java, Intermediate shell scripting
-							(Bash), Intermediate git, Intermediate BASIC, Basic C
-							and C++, Basic R, and Advanced Mac and Linux usage.
+							Intermediate Python, Advanced Java, Intermediate
+							shell scripting (Bash), Intermediate git,
+							Intermediate BASIC, Basic C and C++, Basic R, and
+							Advanced Mac and Linux usage.
 						</p>
 					</div>
 				</section>
@@ -108,6 +113,17 @@ import Hero from "../components/Hero.astro";
 							</p><i><b>UT @ Austin, Texas</b></i>
 						</p>
 						<br />
+						<br />
+						<p>
+							<b
+								>IEEE (Institute of Electrical and Electronics Engineers)-UT</b
+							><p>
+								<i>Student Member</i>
+								<p>
+									Feburary 2025 &nbsp; &ndash; &nbsp; Present
+								</p>
+							</p><i><b>UT @ Austin, Texas</b></i>
+						</p>
 					</div>
 				</section>
 			</section>
