+++ src/pages/coursework.astro
@@ -53,6 +53,7 @@ const fall2023 = [
 ];
 */
 
+/*
 const spring2024 = [
 	{
 		id: 1,
@@ -98,6 +99,7 @@ const spring2024 = [
 			" Students will be given in-class work differing from the homework assigned by the concurrent instructor.",
 	},
 ];
+*/
 
 const summer2024 = [
 	{
@@ -148,10 +150,10 @@ const fall2024 = [
 	},
 	{
 		id: 3,
-		title: "E S 301: Engineering Design/Problem Solving",
+		title: "PHY 303K: Engineering Physics I",
 		description:
-			"Introduction to the scope of engineering, foundations of engineering science, and engineering design." +
-			" May be counted toward the Writing flag requirement.",
+			"A general survey of physics; primarily laws of motion, heat, and wave phenomena." +
+			" Three lecture hours and one discussion hour a week for one semester.",
 	},
 	{
 		id: 4,
@@ -159,16 +161,6 @@ const fall2024 = [
 		description:
 			"Covers fundamentals of probability, combinatorics, discrete and continuous random variables, jointly distributed random variables, and limit theorems. Using probability to introduce fundamentals of statistics, including Bayesian and classical inference." +
 			" The equivalent of four lectures hours a week.",
-	},
-	{
-		id: 5,
-		title: "M 325K: Discrete Mathematics",
-		description:
-			"M325K Discrete Mathematics is the most appropriate upper division mathematics course just after calculus for many of our math majors." +
-			" A few other majors such as computer science or electrical engineering also take this course to develop discrete logic skills." +
-			" In this course students master how to write a formal mathematical proof, how to disprove mathematical statements, and how to find counterexamples." +
-			" As this is a first course that emphasizes understanding and creating proofs, it provides a transition from the problem-solving approach of calculus to the entirely rigorous approach of advanced courses such as M365C or M373K." + 
-			" The number of topics required for coverage has been kept modest so as to allow adequate time for students to develop theorem-proving skills.",
 	},
 ];
 ---
@@ -185,8 +177,8 @@ const fall2024 = [
 			/><br />
 			<!--<h1 style="font-size: large;" ">Fall 2023</h1>
 			<DropdownList dropdownItems={fall2023} /><br/> -->
-			<h1 style="font-size: large;">Spring 2024</h1>
-			<DropdownList dropdownItems={spring2024} /><br />
+			<!-- <h1 style="font-size: large;">Spring 2024</h1>
+			<DropdownList dropdownItems={spring2024} /><br /> -->
 			<h1 style="font-size: large;">Summer 2024</h1>
 			<DropdownList dropdownItems={summer2024} /><br />
 			<h1 style="font-size: large;">Fall 2024</h1>
